Add files via upload
This is currently my main program for which I hope to find the best solution, so my idea works. Ideally, it should download a file, and parse it, outputting the information which could be used for an update. For example, point 6.2. and 8.1.4 from the pdf. could be extracted from this report and compared to the same points in the previous report. If there are no changes, the profile in DMI can be updated.